backend/requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai
python-docx
PyPDF2
Werkzeug # Usually a Flask dependency, but good to list